Refactor classes into src as independent files.
